issues with create_fits
