+++ autogalaxy/analysis/visualizer.py
@@ -4,7 +4,6 @@ from autoconf import conf
 import autoarray.plot as aplt
 
 from autogalaxy.plot.fit_galaxy_plotters import FitGalaxyPlotter
-from autogalaxy.plot.fit_imaging_plotters import FitImagingPlotter
 from autogalaxy.plot.fit_interferometer_plotters import FitInterferometerPlotter
 from autogalaxy.plot.hyper_plotters import HyperPlotter
 
@@ -67,79 +66,6 @@ class Visualizer:
         if should_plot("subplot_dataset"):
 
             imaging_plotter.subplot_imaging()
-
-    def visualize_fit_imaging(self, fit, during_analysis, subfolders="fit_imaging"):
-        def should_plot(name):
-            return plot_setting(section="fit", name=name)
-
-        mat_plot_2d = self.mat_plot_2d_from(subfolders=subfolders)
-
-        fit_imaging_plotter = FitImagingPlotter(
-            fit=fit, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
-        )
-
-        fit_imaging_plotter.figures_2d(
-            image=should_plot("data"),
-            noise_map=should_plot("noise_map"),
-            signal_to_noise_map=should_plot("signal_to_noise_map"),
-            model_image=should_plot("model_data"),
-            residual_map=should_plot("residual_map"),
-            chi_squared_map=should_plot("chi_squared_map"),
-            normalized_residual_map=should_plot("normalized_residual_map"),
-        )
-
-        fit_imaging_plotter.figures_2d_of_galaxies(
-            subtracted_image=should_plot("subtracted_images_of_galaxies"),
-            model_image=should_plot("model_images_of_galaxies"),
-        )
-
-        if should_plot("subplot_fit"):
-            fit_imaging_plotter.subplot_fit_imaging()
-
-        if should_plot("subplots_of_galaxies_fits"):
-            fit_imaging_plotter.subplots_of_galaxies()
-
-        if not during_analysis:
-
-            if should_plot("all_at_end_png"):
-
-                fit_imaging_plotter.figures_2d(
-                    image=True,
-                    noise_map=True,
-                    signal_to_noise_map=True,
-                    model_image=True,
-                    residual_map=True,
-                    normalized_residual_map=True,
-                    chi_squared_map=True,
-                )
-
-                fit_imaging_plotter.figures_2d_of_galaxies(
-                    subtracted_image=True, model_image=True
-                )
-
-            if should_plot("all_at_end_fits"):
-
-                mat_plot_2d = self.mat_plot_2d_from(
-                    subfolders="fit_imaging/fits", format="fits"
-                )
-
-                fit_imaging_plotter = FitImagingPlotter(
-                    fit=fit, mat_plot_2d=mat_plot_2d, include_2d=self.include_2d
-                )
-
-                fit_imaging_plotter.figures_2d(
-                    image=True,
-                    noise_map=True,
-                    signal_to_noise_map=True,
-                    model_image=True,
-                    residual_map=True,
-                    normalized_residual_map=True,
-                    chi_squared_map=True,
-                )
-
-                fit_imaging_plotter.figures_2d_of_galaxies(
-                    subtracted_image=True, model_image=True
-                )
 
     def visualize_interferometer(self, interferometer):
         def should_plot(name):
